Allows strategy used in Brave to avoid allocating on lookup
The Brave tracing library has a modified version of this type, which we
use for tracking in-flight (pending) spans.

https://github.com/openzipkin/brave/blob/master/brave/src/main/java/brave/internal/recorder/PendingSpans.java

I forgot to raise the feature back until now, which completely dodges
classloader issues and the need to use thread locals. If your map key
overrides `equals` in a way that it allows equality with a weak ref of
itself, you don't need any ceremony for lookup-based operations. This
can be a perfectly valid tradeoff for internal code, especially vs
holding up a classloader.